third_party/WebKit/LayoutTests/typedcssom/inlinestyle/width-height.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<div id="testElement"></div>

<script>

var styleMap = testElement.styleMap;

test(function() {
  testElement.style = "";

  styleMap.set('width', new CSSSimpleLength(10, 'px'));
  styleMap.set('height', new CSSSimpleLength(20, 'px'));

  assert_equals(testElement.style.width, '10px');
  assert_equals(testElement.style.height, '20px');
}, "Setting width, height works with a SimpleLength.");

test(function() {
  testElement.style = "";

  styleMap.set('width', new CSSSimpleLength(10, 'percent'));
  styleMap.set('height', new CSSSimpleLength(20, 'percent'));

  assert_equals(testElement.style.width, '10%');
  assert_equals(testElement.style.height, '20%');
}, "Setting width, height works with a Simple percentage.");

test(function() {
  testElement.style = "";

  styleMap.set('width', new CSSCalcLength({px: 5, percent: 6}));
  styleMap.set('height', new CSSCalcLength({px: 7, percent: 8}));

  assert_equals(testElement.style.width, 'calc(6% + 5px)');
  assert_equals(testElement.style.height, 'calc(8% + 7px)');
}, "Setting width, height works with a CalcLength containing a percentage.");

test(function() {
  testElement.style = "width:50em;";

  var width = styleMap.get('width');
  assert_equals(width.constructor, CSSSimpleLength);
  assert_equals(width.value, 50);
  assert_equals(width.type, 'em');
}, "Getting width works with a SimpleLength.");


test(function() {
  testElement.style = "width:60%";

  var width = styleMap.get('width');
  assert_equals(width.constructor, CSSSimpleLength);
  assert_equals(width.value, 60);
  assert_equals(width.type, 'percent');
}, "Getting width works with a simple percentage.");

test(function() {
  testElement.style = "width:calc(60pc + 20%)";

  var width = styleMap.get('width');
  assert_equals(width.constructor, CSSCalcLength);
  assert_equals(width.pc, 60);
  assert_equals(width.percent, 20);
}, "Getting width works with a CalcLength.");

test(function() {
  testElement.style = "height:60vmin;";

  var height = styleMap.get('height');
  assert_equals(height.constructor, CSSSimpleLength);
  assert_equals(height.value, 60);
  assert_equals(height.type, 'vmin');
}, "Getting height works with a SimpleLength.");

test(function() {
  testElement.style = "height:70%;";

  var height = styleMap.get('height');
  assert_equals(height.constructor, CSSSimpleLength);
  assert_equals(height.value, 70);
  assert_equals(height.type, 'percent');
}, "Getting height works with a simple percentage.");

test(function() {
  testElement.style = "height:calc(70% + 4em);";

  var height = styleMap.get('height');
  assert_equals(height.constructor, CSSCalcLength);
  assert_equals(height.percent, 70);
  assert_equals(height.em, 4);
}, "Getting height works with a CalcLength.");

</script>
